backend/requirements.txt
flask
flask-cors
pynacl
supabase
gunicorn
